Move Margins to the root level
