+++ packages/fuselage/src/components/Box/index.js
@@ -95,6 +95,5 @@ Box.defaultProps = {
 
 export { default as AnimatedVisibility } from './AnimatedVisibility';
 export { default as Flex } from './Flex';
-export { default as Margins } from './Margins';
 export { default as Position, PositionAnimated } from './Position';
 export { default as Scrollable } from './Scrollable';
