+++ packages/fuselage/src/components/Margins/spec.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 
-import { Margins } from '../../..';
+import { Margins } from '../..';
 
 it('renders without crashing', () => {
   const div = document.createElement('div');
