+++ packages/fuselage/src/components/Margins/stories.mdx
@@ -2,7 +2,7 @@ import { Meta, Preview, Props, Story } from '@storybook/addon-docs/blocks';
 
 import { Box, Button, ButtonGroup, Flex, Margins, Tile } from '../../..';
 
-<Meta title='Box/Modifiers/Margins' parameters={{ jest: ['Margins/spec'] }} />
+<Meta title='Layout/Margins' parameters={{ jest: ['Margins/spec'] }} />
 
 # Margins
