+++ packages/fuselage/src/components/Select/MultiSelect.js
@@ -1,9 +1,10 @@
 import React, { useState, useRef, useEffect, useCallback } from 'react';
 
-import { AnimatedVisibility, Box, Flex, Margins, Position } from '../Box';
+import { AnimatedVisibility, Box, Flex, Position } from '../Box';
 import { Chip } from '../Chip';
 import { Icon } from '../Icon';
 import { InputBox } from '../InputBox';
+import Margins from '../Margins';
 import { Options, CheckOption, useCursor } from '../Options';
 import { Focus, Addon } from './Select';
