+++ packages/fuselage/src/components/index.js
@@ -17,6 +17,7 @@ export * from './Grid';
 export * from './Icon';
 export * from './InputBox';
 export * from './Label';
+export { default as Margins } from './Margins';
 export * from './Menu';
 export * from './Modal';
 export * from './NumberInput';
